// src/graphql/type.interface.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};
// ...
export type BookingEntity = Node & {
  __typename?: 'BookingEntity';
  addressMoreInfo?: Maybe<Scalars['String']>;
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  latitude: Scalars['Float'];
  longitude: Scalars['Float'];
  mapAddress?: Maybe<Scalars['String']>;
  medias: Array<Media>;
  partner: PartnerEntity;
  partnerId: Scalars['String'];
  problemTexts?: Maybe<Array<Scalars['String']>>;
  problems?: Maybe<Array<CategoryEntity>>;
  quotationAccepted?: Maybe<QuotationEntity>;
  scheduleReason?: Maybe<Scalars['String']>;
  scheduleTime?: Maybe<Scalars['DateTime']>;
  settlementAccepted?: Maybe<SettlementEntity>;
  status: BookingStatusEnum;
  statusDetail?: Maybe<BookingStatusEntity>;
  technician?: Maybe<PartnerEntity>;
  technicianCanReviewUser: Scalars['Boolean'];
  technicianId?: Maybe<Scalars['String']>;
  technicianReviewUser?: Maybe<ReviewEntity>;
  transportDistance: Scalars['Float'];
  transportDuration: Scalars['Float'];
  transportFee: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
  user: UserEntity;
  userCanReviewAgency: Scalars['Boolean'];
  userCanReviewTechnician: Scalars['Boolean'];
  userId: Scalars['String'];
  userReviewAgency?: Maybe<ReviewEntity>;
  userReviewTechnician?: Maybe<ReviewEntity>;
  vehicle: VehicleEntity;
  vehicleId: Scalars['String'];
};

export type BookingStatusEntity = Node & {
  __typename?: 'BookingStatusEntity';
  bookingId: Scalars['String'];
  createdAt: Scalars['DateTime'];
  id: Scalars['ID'];
  note?: Maybe<Scalars['String']>;
  partnerId?: Maybe<Scalars['String']>;
  reasons?: Maybe<Array<CategoryEntity>>;
  scheduleReason?: Maybe<Scalars['String']>;
  scheduleTime?: Maybe<Scalars['DateTime']>;
  status: BookingStatusEnum;
  userId?: Maybe<Scalars['String']>;
};

export enum BookingStatusEnum {
  ASSIGNED_TECHNICIAN = 'ASSIGNED_TECHNICIAN',
  CANCEL = 'CANCEL',
  COMPLETE = 'COMPLETE',
  QUOTATION_ACCEPTED = 'QUOTATION_ACCEPTED',
  QUOTATION_REJECTED = 'QUOTATION_REJECTED',
  QUOTATION_REQUESTED = 'QUOTATION_REQUESTED',
  RESCHEDULED = 'RESCHEDULED',
  SETTLEMENT_ACCEPTED = 'SETTLEMENT_ACCEPTED',
  SETTLEMENT_REJECTED = 'SETTLEMENT_REJECTED',
  SETTLEMENT_REQUESTED = 'SETTLEMENT_REQUESTED',
  TECHNICIAN_ARRIVED = 'TECHNICIAN_ARRIVED',
  TECHNICIAN_ARRIVING = 'TECHNICIAN_ARRIVING',
  WAIT_FOR_CONFIRM = 'WAIT_FOR_CONFIRM',
}
// ...
export type CategoryEntity = Node & {
  __typename?: 'CategoryEntity';
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  name: Scalars['String'];
  type: CategoryTypeEnum;
  updatedAt: Scalars['DateTime'];
};

export enum CategoryTypeEnum {
  CANCEL_MAINTENANCE_REASON = 'CANCEL_MAINTENANCE_REASON',
  CANCEL_ORDER_REASON_BY_PARTNER = 'CANCEL_ORDER_REASON_BY_PARTNER',
  CANCEL_ORDER_REASON_BY_USER = 'CANCEL_ORDER_REASON_BY_USER',
  CANCEL_QUOTATION_REASON = 'CANCEL_QUOTATION_REASON',
  CANCEL_REASON = 'CANCEL_REASON',
  CANCEL_REASON_BY_PARTNER = 'CANCEL_REASON_BY_PARTNER',
  EDUCATION = 'EDUCATION',
  LEVEL = 'LEVEL',
  MANUFACTURER = 'MANUFACTURER',
  MODEL = 'MODEL',
  ORIGIN = 'ORIGIN',
  PART_OF_PRODUCT = 'PART_OF_PRODUCT',
  PROBLEM = 'PROBLEM',
  PRODUCT_UNIT = 'PRODUCT_UNIT',
  QUALIFICATION = 'QUALIFICATION',
  QUOTATION_PRICE_LIST = 'QUOTATION_PRICE_LIST',
  STORE_PRODUCT = 'STORE_PRODUCT',
  VEHICLE_TYPE = 'VEHICLE_TYPE',
}
// ...
export type Expense = {
  __typename?: 'Expense';
  cost?: Maybe<Scalars['Float']>;
  distance?: Maybe<Scalars['Float']>;
  time?: Maybe<Scalars['Float']>;
};
// ...
export enum FileType {
  EXCEL = 'EXCEL',
  FILE = 'FILE',
  IMAGE = 'IMAGE',
  INVOICE = 'INVOICE',
  VIDEO = 'VIDEO',
}
// ...
export type Media = Node & {
  __typename?: 'Media';
  createdAt: Scalars['DateTime'];
  fileSize?: Maybe<Scalars['Int']>;
  fullOriginalUrl?: Maybe<Scalars['String']>;
  fullThumbUrl?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isDeleted: Scalars['Boolean'];
  mimeType?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  originalUrl: Scalars['String'];
  ownerId?: Maybe<Scalars['String']>;
  thumbUrl?: Maybe<Scalars['String']>;
  type: FileType;
  updatedAt: Scalars['DateTime'];
  videoUrl?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export enum OperatingUnitEnum {
  HOURS = 'HOURS',
  KM = 'KM',
}
// ...
export type PartnerEntity = Node & {
  __typename?: 'PartnerEntity';
  addressMoreInfo?: Maybe<Scalars['String']>;
  avatar?: Maybe<Media>;
  avatarId?: Maybe<Scalars['String']>;
  bank: Scalars['String'];
  birthday?: Maybe<Scalars['DateTime']>;
  cardNumber: Scalars['String'];
  citizenId?: Maybe<Scalars['String']>;
  countProduct?: Maybe<Scalars['Int']>;
  countTechnician?: Maybe<Scalars['Int']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  education?: Maybe<CategoryEntity>;
  email?: Maybe<Scalars['String']>;
  expenseInfo?: Maybe<Expense>;
  fullname?: Maybe<Scalars['String']>;
  hotline?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  isAdmin: Scalars['Boolean'];
  isApproved?: Maybe<Scalars['Boolean']>;
  latitude?: Maybe<Scalars['Float']>;
  level?: Maybe<CategoryEntity>;
  longitude?: Maybe<Scalars['Float']>;
  mapAddress?: Maybe<Scalars['String']>;
  menus?: Maybe<Array<PartnerMenuEnum>>;
  parentId?: Maybe<Scalars['String']>;
  parentInfo?: Maybe<PartnerEntity>;
  phone: Scalars['String'];
  qualifications?: Maybe<Array<CategoryEntity>>;
  reviewSummary?: Maybe<ReviewSummary>;
  star: Scalars['Float'];
  starInfo: Array<StarInfo>;
  storeReviewSummary?: Maybe<ReviewSummary>;
  storeStar?: Maybe<Scalars['Float']>;
  storeStarInfo?: Maybe<Array<StarInfo>>;
  suggestionPoint: Scalars['Float'];
  type: PartnerTypeEnum;
  updatedAt: Scalars['DateTime'];
};

export type PartnerForgotPasswordInput = {
  emailOrPhone: Scalars['String'];
};

export type PartnerLoginInput = {
  deviceId?: InputMaybe<Scalars['String']>;
  email?: InputMaybe<Scalars['String']>;
  password: Scalars['String'];
  phone?: InputMaybe<Scalars['String']>;
};

export enum PartnerMenuEnum {
  COURSE = 'COURSE',
  DISCOUNT = 'DISCOUNT',
  DOCUMENT = 'DOCUMENT',
  ORDER = 'ORDER',
  PRODUCT = 'PRODUCT',
  REPAIR = 'REPAIR',
  TECHNICIAN = 'TECHNICIAN',
}
// ...
export enum PartnerTypeEnum {
  AGENCY = 'AGENCY',
  FREELANCER_TECHNICIAN = 'FREELANCER_TECHNICIAN',
  TECHNICIAN = 'TECHNICIAN',
}
// ...
export type QuotationAccessoryEntity = Node & {
  __typename?: 'QuotationAccessoryEntity';
  available: Scalars['Boolean'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  name: Scalars['String'];
  quantity: Scalars['Float'];
  quotationId: Scalars['String'];
  unit: Scalars['String'];
  unitPrice: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
};

export type QuotationAccessoryInput = {
  available: Scalars['Boolean'];
  name: Scalars['String'];
  quantity: Scalars['Float'];
  unit: Scalars['String'];
  unitPrice: Scalars['Float'];
};

export type QuotationAdditionalFeeEntity = Node & {
  __typename?: 'QuotationAdditionalFeeEntity';
  amount: Scalars['Float'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  name: Scalars['String'];
  quotationId: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type QuotationEntity = Node & {
  __typename?: 'QuotationEntity';
  accessories: Array<QuotationAccessoryEntity>;
  additionalFees: Array<QuotationAdditionalFeeEntity>;
  booking: BookingEntity;
  bookingId: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  diagnosisFee: Scalars['Float'];
  diagnosisNote?: Maybe<Scalars['String']>;
  diagnostics: Array<QuotationDiagnosticEntity>;
  estimatedCompleteAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  operatingNumber: Scalars['Float'];
  operatingUnit: OperatingUnitEnum;
  reasons?: Maybe<Array<CategoryEntity>>;
  rejectReasons?: Maybe<Scalars['String']>;
  repairFee: Scalars['Float'];
  status: QuotationStatusEnum;
  technician: PartnerEntity;
  technicianId: Scalars['String'];
  total: Scalars['Float'];
  transportFee: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
  user: UserEntity;
  userId: Scalars['String'];
};
// ...
export enum QuotationStatusEnum {
  ACCEPTED = 'ACCEPTED',
  PENDING = 'PENDING',
  REJECTED = 'REJECTED',
}
// ...
export type ReviewEntity = Node & {
  __typename?: 'ReviewEntity';
  assessorId: Scalars['String'];
  bookingId?: Maybe<Scalars['String']>;
  comment: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  orderId?: Maybe<Scalars['String']>;
  partnerAssessor?: Maybe<PartnerEntity>;
  personEvaluatedId?: Maybe<Scalars['String']>;
  productId?: Maybe<Scalars['String']>;
  star: Scalars['Float'];
  type: ReviewTypeEnum;
  updatedAt: Scalars['DateTime'];
  userAssessor?: Maybe<UserEntity>;
};

export enum ReviewObjectEnum {
  BOOKING = 'BOOKING',
  ORDER = 'ORDER',
}

export type ReviewSummary = {
  __typename?: 'ReviewSummary';
  percent: Scalars['Float'];
  starAverage: Scalars['Float'];
  total: Scalars['Float'];
};
// ...
export type SettlementEntity = Node & {
  __typename?: 'SettlementEntity';
  additionalFees: Array<SettlementAdditionalFeeEntity>;
  booking: BookingEntity;
  bookingId: Scalars['String'];
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  discount?: Maybe<Scalars['Float']>;
  id: Scalars['ID'];
  invoiceId?: Maybe<Scalars['String']>;
  quotation: QuotationEntity;
  quotationId: Scalars['String'];
  status: SettlementStatusEnum;
  technician: PartnerEntity;
  technicianId: Scalars['String'];
  total: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
  user: UserEntity;
  userId: Scalars['String'];
  vatTax?: Maybe<Scalars['Float']>;
};

export enum SettlementStatusEnum {
  ACCEPTED = 'ACCEPTED',
  PENDING = 'PENDING',
  REJECTED = 'REJECTED',
}
// ...
export type StarInfo = {
  __typename?: 'StarInfo';
  star: Scalars['Float'];
  total: Scalars['Float'];
};
// ...
export type UserEntity = Node & {
  __typename?: 'UserEntity';
  address?: Maybe<Scalars['String']>;
  avatar?: Maybe<Media>;
  avatarId?: Maybe<Scalars['String']>;
  birthday?: Maybe<Scalars['DateTime']>;
  certificate?: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  email?: Maybe<Scalars['String']>;
  fullname?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isActive: Scalars['Boolean'];
  numberBooking: Scalars['Float'];
  numberMaintenance: Scalars['Float'];
  numberOrder: Scalars['Float'];
  phone: Scalars['String'];
  star: Scalars['Float'];
  totalBookings: Scalars['Float'];
  totalMaintenanceRequests: Scalars['Float'];
  totalOrders: Scalars['Float'];
  totalPayment: Scalars['Float'];
  totalSpending: Scalars['Float'];
  updatedAt: Scalars['DateTime'];
  userAddress?: Maybe<AddressEntity>;
};
// ...
export type VehicleEntity = Node & {
  __typename?: 'VehicleEntity';
  addressMoreInfo?: Maybe<Scalars['String']>;
  avatar?: Maybe<Media>;
  avatarId?: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  detail?: Maybe<Scalars['String']>;
  hidden: Scalars['Boolean'];
  id: Scalars['ID'];
  latitude: Scalars['Float'];
  longitude: Scalars['Float'];
  manufacturer?: Maybe<CategoryEntity>;
  mapAddress?: Maybe<Scalars['String']>;
  model?: Maybe<CategoryEntity>;
  name: Scalars['String'];
  operatingNumber: Scalars['Float'];
  operatingUnit: OperatingUnitEnum;
  ordinalNumber?: Maybe<Scalars['Float']>;
  origin?: Maybe<CategoryEntity>;
  serialNumber?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
  userId: Scalars['String'];
  vehicleRegistrationPlate?: Maybe<Scalars['String']>;
  vehicleType?: Maybe<CategoryEntity>;
  vinNumber: Scalars['String'];
  yearOfManufacture?: Maybe<Scalars['Float']>;
};